Remove widget classes to separate modules
