+++ chem_textual_ui/chem_ui.py
@@ -11,93 +11,6 @@ from textual.widgets import Button, Header, Footer, Static, Placeholder, Input, 
 import re
 
 from chem_data.data_set import load_data_frame
-def calculate_enthalpy_change(data_frame, reactants: list, products: list):
-    column_name = 'dH'
-    sum_products = 0
-    sum_reactants = 0
-
-    for reactant in reactants:
-        row = reactant[0]
-        coefficient = reactant[1]
-        enthalpy = data_frame._get_value(row, column_name)
-        sum_reactants = sum_reactants + enthalpy * coefficient
-
-    for product in products:
-        row = product[0]
-        coefficient = product[1]
-        enthalpy = data_frame._get_value(row, column_name)
-        sum_products = sum_products + enthalpy * coefficient
-
-    enthalpy_change = sum_products - sum_reactants
-    return enthalpy_change
-
-
-def calculate_free_energy(data_frame, reactants: list, products: list):
-    column_name = 'dG'
-    sum_products = 0
-    sum_reactants = 0
-
-    for reactant in reactants:
-        row = reactant[0]
-        coefficient = reactant[1]
-        enthalpy = data_frame._get_value(row, column_name)
-        sum_reactants = sum_reactants + enthalpy * coefficient
-
-    for product in products:
-        row = product[0]
-        coefficient = product[1]
-        enthalpy = data_frame._get_value(row, column_name)
-        sum_products = sum_products + enthalpy * coefficient
-
-    energy_change = sum_products - sum_reactants
-
-    print("Sum products: ")
-    print(sum_products)
-    print("Sum of reactants:")
-    print(sum_reactants)
-    return energy_change
-
-
-def calculate_entropy_change(data_frame, reactants: list, products: list):
-    column_name = 'dS'
-    sum_products = 0
-    sum_reactants = 0
-
-    for reactant in reactants:
-        row = reactant[0]
-        coefficient = reactant[1]
-        enthalpy = data_frame._get_value(row, column_name)
-        sum_reactants = sum_reactants + enthalpy * coefficient
-
-    for product in products:
-        row = product[0]
-        coefficient = product[1]
-        enthalpy = data_frame._get_value(row, column_name)
-        sum_products = sum_products + enthalpy * coefficient
-
-    # Divide by 1000 because of J to kJ unit conversion
-    entropy_change = (sum_products - sum_reactants) / 1000
-
-    print("Sum products: ")
-    print(sum_products)
-    print("Sum of reactants:")
-    print(sum_reactants)
-    return entropy_change
-
-def extract_chemical_formulas(df, chemical_list):
-    product_formulas = []
-    col_name = 'name'
-    col_state = 'state'
-    col_formula = 'formula'
-    for chemical in chemical_list:
-        row = chemical[0]
-        name = df._get_value(row, col_name)
-        formula = df._get_value(row, col_formula)
-        state = df._get_value(row, col_state)
-        full_formula = f'{formula} {name} {state}'
-        product_formulas.append(full_formula)
-
-    return product_formulas
 
 class ChemApp(App):
     CSS_PATH = "chem_ui.css"
@@ -215,71 +128,5 @@ class MainPanel(Static):
         yield DataWindow()
         yield OutputPanel()
 
-class DataWindow(Static):
-    def compose(self) -> ComposeResult:
-        print("DataWindow compose()")
-        yield TextLog(id="data_log_window")
-
-    def _on_mount(self, event: events.Mount) -> None:
-        log = self.query_one(TextLog)
-        table = self.generate_data_table()
-        log.write(table)
-    def generate_data_table(self):
-        df = load_data_frame()
-        table = Table(title="Chemical Data")
-        table.add_column('index')
-        table.add_column('Name')
-        table.add_column('formula')
-        table.add_column('state')
-        for index, row in df.iterrows():
-            table.add_row(str(index), row["name"],row["formula"],row["state"] )
-        return table
-
-class InstructionsWindow(Static):
-
-    INSTRUCTIONS = '''
-    Instructions\n
-    This app will calculate the changes in enthalpy,
-    free energy, and entropy for a set 
-    reactants and products you select.
-    
-    Use the table to find the chemical you need and enter
-    the index and the coefficient separated by a comma.
-    There is an input box for reactants and one for
-    products. Press the Calculate button to perform
-    the calculations. Clear will clear the screen.
-    '''
-    def compose(self) -> ComposeResult:
-        yield TextLog(id="instruction_log_window")
-
-    def _on_mount(self, event: events.Mount) -> None:
-        instructions_window = self.query_one("#instruction_log_window")
-        #panel = Panel(Text(self.INSTRUCTIONS), expand=False, title="Instructions")
-        text = Text(self.INSTRUCTIONS)
-        instructions_window.write(text)
 
 
-class OutputPanel(Static):
-    def compose(self) -> ComposeResult:
-        yield TextLog(id = "reactants")
-        yield TextLog(id="products")
-        yield TextLog(id= "output")
-
-
-#Container for the input fields and two buttons
-class UserInputArea(Static):
-    def compose(self) -> ComposeResult:
-        yield InputArea()
-        yield Button("Calculate!", id="btn_calculate", variant="success")
-        yield Button("Clear", id="btn_clear", variant="error")
-
-class OutputWindow(Static):
-    def compose(self) -> ComposeResult:
-        yield Placeholder()
-
-class InputArea(Static):
-    def compose(self) -> ComposeResult:
-        yield Label("Reactants")
-        yield Input(placeholder="Example: 3,2 to enter 2 mol butane", id="reactant_input")
-        yield Label("Products")
-        yield Input(placeholder="Example: 0,5 to enter 5 mol methane", id="product_input")
